# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JISHLink workforce management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface Workplace {
  id: string;
  name: string;
  client_name?: string | null;
  address?: string | null;
  latitude?: number | null;
  longitude?: number | null;
}

export interface Employee {
  id: string;
  employee_code: string;
  full_name: string;
  dob?: string | null;
  gender?: string | null;
  aadhar_number?: string | null;
  pan_number?: string | null;
  qualification?: string | null;
  marital_status?: string | null;
  blood_group?: string | null;
  contact_number?: string | null;
  email?: string | null;
  address?: string | null;
  emergency_contact?: string | null;
  nominee_name?: string | null;
  nominee_relation?: string | null;
  designation?: string | null;
  workplace_id?: string | null;
  workplace?: Workplace | null;
  date_of_joining?: string | null;
  date_of_leaving?: string | null;
  employment_status?: string | null;
  employment_type?: string | null;
  reporting_manager_id?: string | null;
  pf_number?: string | null;
  esi_number?: string | null;
  uan_number?: string | null;
  bank_name?: string | null;
  bank_branch?: string | null;
  account_number?: string | null;
  ifsc_code?: string | null;
  driving_license_number?: string | null;
  vehicle_details?: string | null;
  photo_url?: string | null;
  username?: string | null;
  role?: string | null;
  password_changed?: boolean | null;
  created_at?: string | null;
}

export interface AuthLoginResponse {
  token: string;
  employee: Employee;
  passwordChanged: boolean;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface CreateEmployeeRequest {
  full_name: string;
  employee_code: string;
  dob?: string | null;
  gender?: string | null;
  aadhar_number?: string | null;
  pan_number?: string | null;
  qualification?: string | null;
  marital_status?: string | null;
  blood_group?: string | null;
  contact_number?: string | null;
  email?: string | null;
  address?: string | null;
  emergency_contact?: string | null;
  nominee_name?: string | null;
  nominee_relation?: string | null;
  designation?: string | null;
  workplace_id?: string | null;
  date_of_joining?: string | null;
  employment_type?: string | null;
  reporting_manager_id?: string | null;
  pf_number?: string | null;
  esi_number?: string | null;
  uan_number?: string | null;
  bank_name?: string | null;
  bank_branch?: string | null;
  account_number?: string | null;
  ifsc_code?: string | null;
  driving_license_number?: string | null;
  vehicle_details?: string | null;
  photo_url?: string | null;
  username?: string | null;
  password?: string | null;
  role?: string | null;
}

export interface UpdateEmployeeRequest {
  full_name?: string | null;
  dob?: string | null;
  gender?: string | null;
  qualification?: string | null;
  marital_status?: string | null;
  blood_group?: string | null;
  contact_number?: string | null;
  email?: string | null;
  address?: string | null;
  emergency_contact?: string | null;
  nominee_name?: string | null;
  nominee_relation?: string | null;
  designation?: string | null;
  workplace_id?: string | null;
  date_of_joining?: string | null;
  date_of_leaving?: string | null;
  employment_status?: string | null;
  employment_type?: string | null;
  reporting_manager_id?: string | null;
  pf_number?: string | null;
  esi_number?: string | null;
  uan_number?: string | null;
  bank_name?: string | null;
  bank_branch?: string | null;
  account_number?: string | null;
  ifsc_code?: string | null;
  driving_license_number?: string | null;
  vehicle_details?: string | null;
  photo_url?: string | null;
  role?: string | null;
}

export interface CreateWorkplaceRequest {
  name: string;
  client_name?: string | null;
  address?: string | null;
  latitude?: number | null;
  longitude?: number | null;
}

export interface AttendanceLog {
  id: string;
  employee_id: string;
  type: string;
  timestamp: string;
  latitude?: number | null;
  longitude?: number | null;
  resolved_address?: string | null;
}

export interface LogAttendanceRequest {
  type: string;
  latitude?: number | null;
  longitude?: number | null;
  resolved_address?: string | null;
}

export interface TodayStatus {
  logged_in: boolean;
  signed_off: boolean;
  login_time?: string | null;
  signoff_time?: string | null;
  login_address?: string | null;
}

export interface AbsenceNote {
  id: string;
  employee_id: string;
  date: string;
  recruiter_id: string;
  reason: string;
  notes?: string | null;
  created_at: string;
}

export interface CreateAbsenceNoteRequest {
  employee_id: string;
  date: string;
  reason: string;
  notes?: string | null;
}

export type PendingSubmissionSubmittedData = { [key: string]: unknown };

export type PendingSubmissionValidationResults = { [key: string]: unknown } | null;

export interface PendingSubmission {
  id: string;
  source: string;
  submitted_data: PendingSubmissionSubmittedData;
  validation_results?: PendingSubmissionValidationResults;
  status: string;
  admin_remarks?: string | null;
  submitted_at: string;
}

export interface ApproveSubmissionRequest {
  workplace_id: string;
  reporting_manager_id: string;
  username: string;
  password: string;
  designation?: string | null;
}

export interface RejectSubmissionRequest {
  remarks: string;
}

export interface Notification {
  id: string;
  user_id: string;
  message: string;
  read: boolean;
  created_at: string;
}

export interface AdminDashboard {
  total_employees: number;
  active_employees: number;
  pending_approvals: number;
  today_sign_ins: number;
  absent_today: number;
  unread_notifications: number;
}

export interface EmployeeAttendanceStatus {
  employee: Employee;
  logged_in: boolean;
  signed_off: boolean;
  login_time?: string | null;
  signoff_time?: string | null;
  login_address?: string | null;
}

export interface RecruiterDashboard {
  team_count: number;
  signed_in_today: number;
  not_signed_in: number;
  employees: EmployeeAttendanceStatus[];
}

export interface RecruiterStats {
  recruiter: Employee;
  total_employees: number;
  signed_in_today: number;
  not_signed_in_today: number;
  absence_notes_this_month: number;
  reassignments_this_month: number;
}

export type IntakeRequestData = { [key: string]: unknown };

export interface IntakeRequest {
  source: string;
  submittedAt?: string;
  data: IntakeRequestData;
}

export type ImportEmployeesRequestRowsItem = { [key: string]: unknown };

export interface ImportEmployeesRequest {
  rows: ImportEmployeesRequestRowsItem[];
}

export type ImportResultErrorsItem = { [key: string]: unknown };

export interface ImportResult {
  success_count: number;
  error_count: number;
  errors: ImportResultErrorsItem[];
}

export interface ReassignRequest {
  workplace_id: string;
}

export type ListEmployeesParams = {
search?: string;
status?: string;
workplace_id?: string;
};

export type GetMyAttendanceParams = {
days?: number;
};

export type ListSubmissionsParams = {
status?: string;
};