# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JISHLink workforce management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: workplace_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /attendance:
    post:
      operationId: logAttendance
      tags: [attendance]
      summary: Log attendance (login or signoff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogAttendanceRequest"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceLog"
  /attendance/my:
    get:
      operationId: getMyAttendance
      tags: [attendance]
      summary: Get my attendance logs
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: My attendance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceLog"
...
  /employees/import:
    post:
      operationId: importEmployees
      tags: [employees]
      summary: Bulk import employees from parsed Excel rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportEmployeesRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /employees/{id}/reassign:
    post:
      operationId: reassignEmployee
      tags: [employees]
      summary: Reassign employee to a different workplace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReassignRequest"
      responses:
        "200":
          description: Reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
        passwordChanged:
          type: boolean
      required: [token, employee, passwordChanged]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    Employee:
      type: object
      properties:
        id:
          type: string
        employee_code:
          type: string
        full_name:
          type: string
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        aadhar_number:
          type: string
          nullable: true
        pan_number:
          type: string
          nullable: true
        qualification:
          type: string
          nullable: true
        marital_status:
          type: string
          nullable: true
        blood_group:
          type: string
          nullable: true
        contact_number:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        emergency_contact:
          type: string
          nullable: true
        nominee_name:
          type: string
          nullable: true
        nominee_relation:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        workplace_id:
          type: string
          nullable: true
        workplace:
          $ref: "#/components/schemas/Workplace"
          nullable: true
        date_of_joining:
          type: string
          nullable: true
        date_of_leaving:
          type: string
          nullable: true
        employment_status:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        reporting_manager_id:
          type: string
          nullable: true
        pf_number:
          type: string
          nullable: true
        esi_number:
          type: string
          nullable: true
        uan_number:
          type: string
          nullable: true
        bank_name:
          type: string
          nullable: true
        bank_branch:
          type: string
          nullable: true
        account_number:
          type: string
          nullable: true
        ifsc_code:
          type: string
          nullable: true
        driving_license_number:
          type: string
          nullable: true
        vehicle_details:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        password_changed:
          type: boolean
          nullable: true
        created_at:
          type: string
          nullable: true
      required: [id, employee_code, full_name]
    CreateEmployeeRequest:
      type: object
      properties:
        full_name:
          type: string
        employee_code:
          type: string
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        aadhar_number:
          type: string
          nullable: true
        pan_number:
          type: string
          nullable: true
        qualification:
          type: string
          nullable: true
        marital_status:
          type: string
          nullable: true
        blood_group:
          type: string
          nullable: true
        contact_number:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        emergency_contact:
          type: string
          nullable: true
        nominee_name:
          type: string
          nullable: true
        nominee_relation:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        workplace_id:
          type: string
          nullable: true
        date_of_joining:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        reporting_manager_id:
          type: string
          nullable: true
        pf_number:
          type: string
          nullable: true
        esi_number:
          type: string
          nullable: true
        uan_number:
          type: string
          nullable: true
        bank_name:
          type: string
          nullable: true
        bank_branch:
          type: string
          nullable: true
        account_number:
          type: string
          nullable: true
        ifsc_code:
          type: string
          nullable: true
        driving_license_number:
          type: string
          nullable: true
        vehicle_details:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
      required: [full_name, employee_code]
    UpdateEmployeeRequest:
      type: object
      properties:
        full_name:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        qualification:
          type: string
          nullable: true
        marital_status:
          type: string
          nullable: true
        blood_group:
          type: string
          nullable: true
        contact_number:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        emergency_contact:
          type: string
          nullable: true
        nominee_name:
          type: string
          nullable: true
        nominee_relation:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        workplace_id:
          type: string
          nullable: true
        date_of_joining:
          type: string
          nullable: true
        date_of_leaving:
          type: string
          nullable: true
        employment_status:
          type: string
          nullable: true
        employment_type:
          type: string
          nullable: true
        reporting_manager_id:
          type: string
          nullable: true
        pf_number:
          type: string
          nullable: true
        esi_number:
          type: string
          nullable: true
        uan_number:
          type: string
          nullable: true
        bank_name:
          type: string
          nullable: true
        bank_branch:
          type: string
          nullable: true
        account_number:
          type: string
          nullable: true
        ifsc_code:
          type: string
          nullable: true
        driving_license_number:
          type: string
          nullable: true
        vehicle_details:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
    Workplace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        client_name:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required: [id, name]
...
    AttendanceLog:
      type: object
      properties:
        id:
          type: string
        employee_id:
          type: string
        type:
          type: string
        timestamp:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        resolved_address:
          type: string
          nullable: true
      required: [id, employee_id, type, timestamp]
    LogAttendanceRequest:
      type: object
      properties:
        type:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        resolved_address:
          type: string
          nullable: true
      required: [type]
...
    ImportEmployeesRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [rows]
    ImportResult:
      type: object
      properties:
        success_count:
          type: integer
        error_count:
          type: integer
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [success_count, error_count, errors]
    ReassignRequest:
      type: object
      properties:
        workplace_id:
          type: string
      required: [workplace_id]