# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JISHLink workforce management API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginRequest {
  username: string;
  password: string;
}